+++ domain/src/main/java/com/karthik/pro/engr/github/api/domain/calculator/PercentageCalculator.kt
@@ -2,8 +2,8 @@ package com.karthik.pro.engr.github.api.domain.calculator
 
 object PercentageCalculator {
 
-    fun calculate(bytes: Int, total: Int): Float {
-        if (total == 0) return 0f
+    fun calculate(bytes: Long, total: Long): Float {
+        if (total == 0L) return 0f
         return (bytes.toFloat() / total) * 100f
     }
 }
